feat: creating table sections
