+++ src/pages/dashboard.jsx
@@ -1,8 +1,8 @@
 import { Input } from "../components/common/input";
 import { Select } from "../components/common/select";
 
-import { TableHeader } from "../components/table/header";
 import { DashboardHeader } from "../containers/dashboardHeader";
+import { TableHeader, TableBody } from "../containers/tableSections";
 
 import { FaSearch } from "react-icons/fa";
 
@@ -10,7 +10,7 @@ import "../styles/table.css";
 
 export const DashBoard = () => {
    return (
-      <section className="flex flex-col gap-10 w-screen h-screen px-96 bg-gray-50 overflow-y-auto">
+      <section className="flex flex-col gap-10 w-screen h-screen px-72 bg-gray-50 overflow-y-auto">
          <DashboardHeader />
 
          <div className="flex flex-col gap-8 w-full h-screen">
@@ -27,8 +27,9 @@ export const DashBoard = () => {
                />
             </header>
 
-            <table className="flex content-start w-full flex-grow">
+            <table className="flex flex-col gap-2 content-start w-full flex-grow">
                <TableHeader />
+               <TableBody />
             </table>
          </div>
       </section>
